+++ controllers/userControllers.js
@@ -1,0 +1,83 @@
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+const User = require("../models/users");
+
+const saltRounds = 10;
+
+async function registerUser(req, res, next) {
+  try {
+    const { error } = registrationSchema.validate(req.body);
+    if (error) {
+      return res.status(400).json({ message: error.details[0].message });
+    }
+
+    const { email, password } = req.body;
+
+    const existingUser = await User.findOne({ email });
+    if (existingUser) {
+      return res.status(409).json({ message: "Email in use" });
+    }
+
+    const passwordHash = await bcrypt.hash(password, saltRounds);
+
+    const newUser = await User.create({ email, password: passwordHash });
+
+    res.status(201).json({
+      user: {
+        email: newUser.email,
+        subscription: "starter",
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+}
+
+async function loginUser(req, res, next) {
+  try {
+    const { email, password } = req.body;
+
+    const user = await User.findOne({ email }).exec();
+
+    if (!user) {
+      return res
+        .status(401)
+        .send({ message: "Email or password is incorrect" });
+    }
+    const isMatch = await bcrypt.compare(password, user.password);
+
+    if (!isMatch) {
+      return res
+        .status(401)
+        .send({ message: "Email or password is incorrect" });
+    }
+
+    const token = jwt.sign(
+      {
+        id: user._id,
+        name: user.name,
+      },
+      process.env.JWT_TOKEN,
+      {
+        expiresIn: 3600,
+      }
+    );
+
+    await User.findByIdAndUpdate(user._id, { token }).exec();
+
+    res.send({ token });
+  } catch (error) {
+    next(error);
+  }
+}
+
+async function logoutUser(req, res, next) {
+  try {
+    await User.findByIdAndUpdate(req.user._id, { token: null });
+    res.status(204).end();
+  } catch (error) {
+    next(error);
+  }
+}
+
+module.exports = { registerUser, loginUser, logoutUser };
