+++ controllers/users.js
@@ -1,95 +1,17 @@
 const express = require("express");
-
 const Joi = require("joi");
-
 const User = require("../models/users");
-
-const bcrypt = require("bcrypt");
-
-const saltRounds = 10;
+const {
+  registerUser,
+  loginUser,
+  logoutUser,
+} = require("../controllers/userControllers");
 
 const router = express.Router();
 
-const registrationSchema = Joi.object({
-  email: Joi.string().email().required(),
-  password: Joi.string().min(6).required(),
-});
 
-router.post("/register", async (req, res, next) => {
-  try {
-    const { error } = registrationSchema.validate(req.body);
-    if (error) {
-      return res.status(400).json({ message: error.details[0].message });
-    }
+router.post("/register", registerUser);
+router.post("/login", loginUser);
+router.post("/logout", logoutUser);
 
-    const { email, password } = req.body;
-
-    const existingUser = await User.findOne({ email });
-    if (existingUser) {
-      return res.status(409).json({ message: "Email in use" });
-    }
-
-    const passwordHash = bcrypt.hash(password, saltRounds);
-
-    const newUser = await User.create({ email, password });
-
-    res.status(201).json({
-      user: {
-        email: newUser.email,
-        subscription: "starter",
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-async function login(req, res, next) {
-  const { email, password } = req.body;
-
-  try {
-    const user = await User.findOne({ email }).exec();
-
-    if (user === null) {
-      console.log("Email");
-      return res
-        .status(401)
-        .send({ message: "Email or password is incorrect" });
-    }
-    const isMatch = await bcrypt.compare(password, user.password);
-
-    if (isMatch !== true) {
-      return res
-        .status(401)
-        .send({ message: "Email or password is incorrect" });
-    }
-
-    const token = jwt.sign(
-      {
-        id: user._id,
-        name: user.name,
-      },
-      process.env.JWT_TOKEN,
-      {
-        expiresIn: 3600,
-      }
-    );
-
-    await User.findByIdAndUpdate(user._id, { token }).exec();
-
-    res.send({ token });
-  } catch (error) {
-    next(error);
-  }
-}
-
-async function logout(req, res, next) {
-  try {
-    await User.findByIdAndUpdate(req.user._id, { token: null }).exec();
-    res.status(204).end();
-  } catch (error) {
-    next(error);
-  }
-}
-
-module.exports = { router, login, logout };
+module.exports = router;
