+++ utils/validation/userValidationSchemas.js
@@ -1,0 +1,16 @@
+const Joi = require("joi");
+
+const registrationSchema = Joi.object({
+  email: Joi.string().email().required(),
+  password: Joi.string().min(6).required(),
+});
+
+const loginSchema = Joi.object({
+  email: Joi.string().email().required(),
+  password: Joi.string().min(6).required(),
+});
+
+module.exports = {
+  registrationSchema,
+  loginSchema,
+};
